requirements.txt
Flask==0.9
Flask-KVSession==0.3.2
Flask-WTF==0.8
PIL==1.1.7
simple-pbkdf2==1.0
